Refactor PR Details page components and state
